at this point we're making great progress
I guess the bigger theme is to figure out to which extent it's possible to template everything. For example title positions vary a great deal based on the plot size, so it might be smarter to not even try in terms of the x and y position. At least plots should be broken to several types eg. single plot, subplot, x/y axis , etc.